backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic
python-multipart
python-dotenv
groq
pytest
requests